[DISCO-4333] feat(fleece): part 1 - refactor to make log models reusable
