+++ merino-common/merino_common/models/__init__.py
@@ -1,0 +1,1 @@
+"""Shared data models for Merino services."""
